Start client prototype and break apart server and client
Summary:
Add the start of the client-side work. It's pretty rough, but it has most of the
basic I/O to capture the info we want. The server and client codebases are
broken apart because brunch was being a pain to get working properly. It was
trying to do way more than we need right now.

Test Plan:
- Build and install the client to be served by the server (`cd client && make deps && make install`)
- Build and start the server (`cd server && mix do deps.get, phoenix.server`)
- Go to `localhost:4000` and you should see the client

Reviewers: sethetter

Reviewed By: sethetter

Maniphest Tasks: T20

Differential Revision: https://phabricator.opengovfoundation.org/D6