Refactor nixops files
Summary:
Break apart the logical and physical specifications a bit better, i.e., actually
use NixOS modules correctly. It's not perfect as `config/resolutionizer.nix`
still references `deployment.*` keys, but if we want to support a plain NixOS
module for the resolutionizer we aren't too far off, would take a bit more
refactoring. But our focus for the module is in the context of nixops for now,
so it's fine.

Breaking this stuff apart makes it easier to deploy multiple instances and to
different targets (e.g., a local VM, container, a VPS, etc.).

Test Plan: Look at it.

Reviewers: sethetter

Reviewed By: sethetter

Differential Revision: https://phabricator.opengovfoundation.org/D25