Use Georgia for resolution font
Summary:
Also disable the fontconfig-ultimate settings. They can either be
disabled or the substitutions need to switch to either "ms" or "combi"
from it's default "free" in order to use the Microsoft fonts. Switching
the substitutes is fine, but disabling all together seems to give
//slightly// better looking rendering for Georgia, so just disabled it for
now.

The `font-weight: 600` setting was making rendering weird, probably because
there is no 600 weight font available, so just use normal weight fonts for
normal things. The other option is to make the whole thing bold, which
seems terrible, or to embed the fonts with the proper font weights we need
into the CSS.

Test Plan: Generate a PDF with the tool.

Reviewers: sethetter

Reviewed By: sethetter

Maniphest Tasks: T56

Differential Revision: https://phabricator.opengovfoundation.org/D21