Use Georgia for resolution font
Summary:
Also disable the fontconfig-ultimate settings. They can either be
disabled or the substitutions need to switch to either "ms" or "combi"
from it's default "free" in order to use the Microsoft fonts. Switching
the substitutes is fine, but disabling all together seems to give
*slightly* better looking rendering for Georgia, so just disabled it for
now.

Test Plan:
Generate a PDF with the tool.

Reviewers: sethetter

Subscribers: