Rework the sponsor selector UI
Summary:
Get rid of the confusing selector with plus button and instead have
a simple button which adds a new sponsor selector for them to then
choose from. The previous approach failed to make enough distinction
between the selector to choose which sponsor to add and the selectors to
change existing sponsors.

The new approach also improves how we disallow duplicate sponsors, with
each sponsor selector excluding any sponsors that are selected
elsewhere. So the UI no longer even exposes the possibility of selecting
duplicate sponsors.

Other things:
- Include a sponsor selector by default since all resolutions will
  have a least one sponsor.
- Misc. UI fixes.

Test Plan: Play with it.

Reviewers: sethetter

Reviewed By: sethetter

Differential Revision: https://phabricator.opengovfoundation.org/D24