Introduce dialyxir and elixir type annotations
Summary:
* Included dialyxr to bring dialyzer static analysis to our project.
* Put in type annotations for the PDF modules.

Currently, dialyzer does not produce a clean report with a base phoenix project,
so the dialyzer run has not been included as part of the build process. [[ https://github.com/phoenixframework/phoenix/issues/1872 |
These issues are known and being fixed ]] so we will keep an eye on them and
introduce dialyzer to the build process when it's appropriate.

Ref T58

Test Plan:
* `cd server && mix deps.get`
* `mix dialyzer.plt`
* `mix dialyzer`
* You will see a bunch of errors, capture it and there should be 29 lines (1 blank)

Reviewers: doshitan

Reviewed By: doshitan

Maniphest Tasks: T58

Differential Revision: https://phabricator.opengovfoundation.org/D28