Fix local database creation
Was a bit puzzled as to why I needed to use `sudo` in the first place,
but it seemed to work so I just moved on. Thinking about it more, the
Postgres environment variables for the application were causing the
issues. Things make more sense now. For initial database setup we
obviously don't want to try to connect as the app or to the app database
since neither exist yet.