Format sponsor names in parchment output
Summary:
Sponsor names in the parchment resolution output should be..

* Grouped by title (Aldermen, Mayor, Clerk)
* In the "First Last" format
* Leave out the ward number

I would love for this to be a bit more clean in the future, having this giant
function in the template is not pretty. However, we'll need to think through the
ability to group data formatting functions with different templates.

One other thing: I'm not sure if the order between Aldermen, Mayor and Clerk
matters. I've posted this in the Chicago channel in Slack so we remember to ask.

Closes T80

Test Plan:
```
server/test/integration/bin/docgentest &&
  xdg-open server/test/integration/_docgentest-results
```

Reviewers: doshitan

Maniphest Tasks: T80

Differential Revision: https://phabricator.opengovfoundation.org/D44